Size

+++ crates/swc_ecma_minifier/src/analyzer/mod.rs
@@ -78,28 +78,28 @@ pub(crate) struct VarUsageInfo {
     pub inline_prevented: bool,
 
     /// The number of direct reference to this identifier.
-    pub ref_count: usize,
+    pub ref_count: u32,
 
     /// `true` if a variable is conditionally initialized.
     pub cond_init: bool,
 
     /// `false` if it's only used.
     pub declared: bool,
-    pub declared_count: usize,
+    pub declared_count: u32,
 
     /// `true` if the enclosing function defines this variable as a parameter.
     pub declared_as_fn_param: bool,
 
     pub declared_as_fn_expr: bool,
 
-    pub assign_count: usize,
-    pub mutation_by_call_count: usize,
+    pub assign_count: u32,
+    pub mutation_by_call_count: u32,
 
     /// The number of direct and indirect reference to this identifier.
     /// ## Things to note
     ///
     /// - Update is counted as usage
-    pub usage_count: usize,
+    pub usage_count: u32,
 
     /// The variable itself is modified.
     pub reassigned_with_assignment: bool,
@@ -109,7 +109,7 @@ pub(crate) struct VarUsageInfo {
 
     pub has_property_access: bool,
     pub has_property_mutation: bool,
-    pub accessed_props: AHashMap<JsWord, usize>,
+    pub accessed_props: AHashMap<JsWord, u32>,
 
     pub exported: bool,
     /// True if used **above** the declaration or in init. (Not eval order).
